# microservices/auth-service/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: '0.1.0'
  title: Users API
  description: An API to hand users with authentication/authorization
host: localhost:8080
basePath: /api
schemes:
  - http

paths:
  /connection:
    post:
      description: get jwt token
      operationId: login
      consumes:
        - application/json
      parameters:
        - name: credentials
          in: body
          required: true
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        '200':
          description: successfuly loged in
          schema:
            type: object
            $ref: '#/definitions/Token'
        '401':
          description: unvalid credentials

  /users:
    post:
      description: create a user
      operationId: createUser
      consumes:
        - application/json
      parameters:
        - name: user
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: user created successfully
        '409':
          description: email already used
    get:
      description: get the list of all users
      operationId: getUsers
      produces:
        - application/json
      responses:
        '200':
          description: users retrieved successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/UserDTO'

  /users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        type: string
    delete:
      description: delete a user
      operationId: deleteUser
      responses:
        '204':
          description: user deleted successfully
        '404':
          description: user not found
    patch:
      description: modify the password of a user
      operationId: changePassword
      consumes:
        - text/plain
      parameters:
        - in: body
          name: password
          required: true
          schema:
            type: string
      responses:
        '204':
          description: password modified successfully
        '404':
          description: user not found
definitions:
  User:
    type: object
    properties:
      email:
        type: string
      isAdmin:
        type: boolean
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
  UserDTO:
    type: object
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
  Credentials:
    type: object
    properties:
      email:
        type: string
      pwHash:
        type: string
  Token:
    type: object
    properties:
      token:
        type: string